New update
When adding an impact player, the software will now remove player from home/away substitute list. And undoing NEW BATSMAN will place the player back to the substitution list.
If you now find a player in batting card who is also in the substitution list, then ignore them. Please test this scoring and pass it on.